requirements.txt
appdirs==1.4.3
django-environ==0.4.5
django-rest-auth==0.9.5
dj-static==0.0.6
Django==3.2.5
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
gunicorn==20.1.0
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
